requirements.txt
python-telegram-bot
requests
schedule
dotenv
telebot
